Add files via upload
$report = "C:\Reports\SystemReport_$(Get-Date -Format 'yyyyMMdd').txt":

This line creates a variable named $report and assigns it a file path where the system report will be saved. The file path includes the directory C:\Reports\ and a filename with the format SystemReport_<current date>.txt, where the <current date> is obtained using the Get-Date cmdlet with the format yyyyMMdd, representing the year, month, and day.

Get-WmiObject -Class Win32_ComputerSystem | Out-File -FilePath $report:

This line retrieves system information related to the computer system using the Get-WmiObject cmdlet with the -Class parameter set to Win32_ComputerSystem. This class provides information about the computer system, such as the manufacturer, model, and system type.
The retrieved information is then piped (|) to the Out-File cmdlet, which saves the output to the file specified by the $report variable. Since this is the first command to write to the file, it creates a new file or overwrites an existing one.

Get-WmiObject -Class Win32_Processor | Out-File -FilePath $report -Append:

This line retrieves information about the processor (CPU) using the Win32_Processor class.
Similar to the previous line, the retrieved information is piped to the Out-File cmdlet with the -Append parameter. This parameter appends the output to the existing file specified by the $report variable, rather than overwriting it.

Get-WmiObject -Class Win32_LogicalDisk | Out-File -FilePath $report -Append:

This line retrieves information about the logical disks (hard drives) using the Win32_LogicalDisk class.
Again, the retrieved information is piped to the Out-File cmdlet with the -Append parameter to append the output to the existing file specified by the $report variable.