Add files via upload
Get-WmiObject -Class Win32_Product: This cmdlet retrieves instances of the Win32_Product class, which contains information about installed software products on the system.

ForEach-Object { ... }: This is a loop construct that iterates over each object returned by the Get-WmiObject cmdlet.

$name = $_.Name: Within the loop, $_ represents the current object being processed. In this case, it's an instance of the Win32_Product class. $_.Name retrieves the name of the software product and assigns it to the $name variable.

$version = $_.Version: Similarly, $_.Version retrieves the version of the software product and assigns it to the $version variable.

Write-Host "$name - Version $version": This line prints out the name and version of each software product to the console. It uses string interpolation to include the values of the $name and $version variables within the string.